Separate Classes into different files
Rather than having a monolithic file, separate classes and components
into separate files
