+++ mbget/impl/github_downloader.py
@@ -1,0 +1,61 @@
+import re
+
+import requests
+from github import Github, GitReleaseAsset, GitRelease
+
+from impl.barrel_asset import BarrelAsset
+from impl.errors import Error
+from impl.dependency import Dependency
+
+
+class GithubDownloader(object):
+    BARREL_FILE = re.compile(r'^.+\.barrel$')
+
+    def __init__(self, token: str = None):
+        self.__token = token
+
+        if token is not None:
+            self.__github = Github(login_or_token=token)
+        else:
+            self.__github = Github()
+
+    def download_barrel(self, dependency: Dependency) -> BarrelAsset:
+        release = self.__find_release(dependency)
+        asset = self.__get_barrel_asset(release)
+
+        print("Downloading asset {asset} from release {rel}".format(asset=asset.name,
+                                                                    rel=release.tag_name))
+
+        content = self.__request_barrel_content(asset)
+        return BarrelAsset(asset.name, release.tag_name, content)
+
+    def __request_barrel_content(self, asset: GitReleaseAsset) -> bytes:
+        # Found a barrel file, Download it.
+        headers = {'Accept': 'application/octet-stream'}
+
+        if self.__token is not None:
+            headers['Authorization'] = 'token {token}'.format(token=self.__token)
+
+        req = requests.get(asset.url, headers=headers)
+        return req.content
+
+    def __get_barrel_asset(self, release: GitRelease) -> GitReleaseAsset:
+        # Matching tag download the barrels
+        for asset in release.get_assets():
+            if self.BARREL_FILE.match(asset.name) is not None:
+                return asset
+
+        raise Error("No barrel asset found in release: {rel}".format(rel=release.tag_name))
+
+    def __find_release(self, dependency: Dependency) -> GitRelease:
+        repo = self.__github.get_repo(dependency.repo)
+
+        # Search the releases for a version that we can use
+        for release in repo.get_releases():
+            if not dependency.version.matches(release.tag_name):
+                # Version does not match our requirement
+                continue
+            return release
+
+        raise Error("Unable to find matching version {version}".format(
+            version=dependency.version))
